Add tox.ini for local testing.
Pin setuptools and zc.buildout to the same versions in all environments.
This avoids buildout upgrading and restarting itself, which gives an error in tox because the buildout config is no longer found.
I tried using zc.buildout 3 and latest setuptools, but latest setuptools is not installable on Python 3.6.
So stick to zc.buildout 2 for the moment.
We could also pin different versions for Python 3.6 only: that works in both requirements.txt and in buildout configs.